feat: split SidebarItem implemenation in different files
