+++ packages/picasso/src/SidebarItem/SidebarItemHeader.tsx
@@ -1,0 +1,77 @@
+import { makeStyles, Theme } from '@material-ui/core/styles'
+import cx from 'classnames'
+import React, { forwardRef, useCallback } from 'react'
+
+import MenuItem from '../MenuItem'
+import SidebarItemContent from '../SidebarItemContent'
+import styles from './styles'
+import { Props } from './types'
+
+const useStyles = makeStyles<Theme>(styles, {
+  name: 'PicassoSidebarItem'
+})
+
+export const SidebarItemHeader = forwardRef<HTMLElement, Props>(
+  function SidebarItemHeader(props: Props, ref) {
+    const {
+      className,
+      disabled,
+      menu,
+      selected,
+      compact,
+      variant = 'light',
+      onClick,
+      collapsible,
+      // these props are being destructured only for the purpose of excluding them from `...rest`
+      /* eslint-disable @typescript-eslint/no-unused-vars */
+      badge,
+      testIds,
+      icon,
+      isExpanded,
+      expand,
+      index,
+      /* eslint-enable */
+      ...rest
+    } = props
+
+    const classes = useStyles()
+
+    const hasMenu = menu != null
+
+    const handleMenuItemClick = useCallback(
+      (event: React.MouseEvent<HTMLElement, MouseEvent>) => {
+        if (!hasMenu) {
+          onClick?.(event)
+        }
+      },
+      [hasMenu, onClick]
+    )
+
+    return (
+      <MenuItem
+        {...rest}
+        classes={classes}
+        ref={ref}
+        className={cx(
+          classes.root,
+          classes.noWrap,
+          classes.roundedBorder,
+          classes[variant],
+          {
+            [classes.compact]: compact,
+            [classes.selected]: !hasMenu && selected,
+            [classes.collapsible]: hasMenu && collapsible
+          },
+          className
+        )}
+        onClick={handleMenuItemClick}
+        selected={!hasMenu && selected}
+        disabled={disabled}
+        variant={variant}
+        nonSelectable
+      >
+        <SidebarItemContent {...props} />
+      </MenuItem>
+    )
+  }
+)
